+++ backend/src/routes/v1/auth.routes.ts
@@ -1,11 +1,11 @@
 import { Router } from 'express'
 
-import { loginController, logoutController, meController, refreshController } from '../../controllers/auth.controller'
+import { changeOwnPasswordController, loginController, logoutController, meController, refreshController } from '../../controllers/auth.controller'
 import { asyncHandler } from '../../middlewares/async-handler'
 import { requireAuthentication } from '../../middlewares/auth.middleware'
 import { authLoginRateLimit, authRefreshRateLimit } from '../../middlewares/auth-rate-limit.middleware'
 import { validateRequest } from '../../middlewares/validate.middleware'
-import { loginBodySchema } from '../../validators/auth.validator'
+import { changeOwnPasswordBodySchema, loginBodySchema } from '../../validators/auth.validator'
 
 const authRouter = Router()
 
@@ -13,5 +13,6 @@ authRouter.post('/login', authLoginRateLimit, validateRequest({ body: loginBodyS
 authRouter.post('/logout', asyncHandler(logoutController))
 authRouter.post('/refresh', authRefreshRateLimit, asyncHandler(refreshController))
 authRouter.get('/me', requireAuthentication, asyncHandler(meController))
+authRouter.patch('/password', requireAuthentication, validateRequest({ body: changeOwnPasswordBodySchema }), asyncHandler(changeOwnPasswordController))
 
 export { authRouter }